+++ utimes.c
@@ -6,16 +6,42 @@
 
 #include <stdio.h>
 #include <errno.h>
+#include <limits.h>
 #include <fcntl.h>
+#include <dlfcn.h>
 
 #include <sys/time.h>
 
 #include "iamroot.h"
 
+__attribute__((visibility("hidden")))
+int next_utimes(const char *path, const struct timeval times[2])
+{
+	int (*sym)(const char *, const struct timeval[2]);
+	int ret;
+
+	sym = dlsym(RTLD_NEXT, "utimes");
+	if (!sym)
+		return __dl_set_errno(ENOSYS, -1);
+
+	ret = sym(path, times);
+	if (ret == -1)
+		__pathperror(path, __func__);
+
+	return ret;
+}
+
 int utimes(const char *path, const struct timeval times[2])
 {
-	__debug("%s(path: '%s', ...)\n", __func__, path);
+	char buf[PATH_MAX];
+	ssize_t siz;
 
-	/* Forward to another function */
-	return futimesat(AT_FDCWD, path, times);
+	siz = path_resolution(AT_FDCWD, path, buf, sizeof(buf),
+			      AT_SYMLINK_NOFOLLOW);
+	if (siz == -1)
+		return __path_resolution_perror(path, -1);
+
+	__debug("%s(path: '%s' -> '%s', ...)\n", __func__, path, buf);
+
+	return next_utimes(buf, times);
 }